Remove xmlx package and associated test files
- Deleted the entire xmlx package including all its source files and test files.
- This includes the main node handling logic, XML parsing, and utility functions.
- Removed test cases that were validating the functionality of the xmlx package.

+++ internal/adminapi/sessions_test.go
@@ -561,7 +561,7 @@ func TestSessionsEdgeCases(t *testing.T) {
 	t.Run("Query after deletion", func(t *testing.T) {
 		db.Exec("DELETE FROM radius_online")
 		session := createTestOnlineSession(db, "delete-test", "192.168.40.1", "10.4.0.1")
-		sessionID := strconv.FormatInt(int64(session.ID), 10)
+		sessionID := strconv.FormatInt(session.ID, 10)
 
 		// Delete first
 		e := setupTestEcho()
